requirements.txt
flask==3.1.0
flask-restx==1.3.0
flask_cors
pymongo==4.10.0
werkzeug==3.1.3
gunicorn
python-dotenv
flask-jwt-extended
pyyaml
pytest
mongomock
pytest-selenium
selenium
webdriver-manager
bcrypt
bandit
checkov
flask_talisman
google-generativeai 